Move Generation and GenTesting to (ideally) their final resting places
Change imports as necessary to reflect new placement
