+++ package.yaml
@@ -64,6 +64,7 @@ library:
     - Language.Pads.CodeGen
     - Language.Pads.CoreBaseTypes
     - Language.Pads.Errors
+    - Language.Pads.Generation
     - Language.Pads.Generic
     - Language.Pads.GenPretty
     - Language.Pads.MetaData
@@ -82,7 +83,6 @@ library:
     - Language.Pads.Library.BinaryUtilities
     - Language.Pads.Library.LittleEndian
     - Language.Pads.Library.Native
-    - Language.Pads.DataGen.Generation
 
 tests:
   examples:
@@ -108,11 +108,11 @@ tests:
   # examples/Main.hs: (should your module be named something other than Main,
   # the -main-is flag will be necessary):
   gen:
-    main: src/Language/Pads/DataGen/GenTesting.hs
+    main: examples/GenTesting.hs
     ghc-options:
     - -threaded
     - -rtsopts
     - -with-rtsopts=-N
-    - -main-is Language.Pads.DataGen.GenTesting
+    - -main-is GenTesting
     dependencies:
     - pads-haskell
