+++ src/Language/Pads/Padsc.hs
@@ -31,7 +31,7 @@ module Language.Pads.Padsc (
      module Language.Pads.BaseTypes,
      module Language.Pads.Pretty,
      module Language.Pads.PadsPrinter,
-     module Language.Pads.DataGen.Generation,
+     module Language.Pads.Generation,
      module Data.Data,
      module Data.List,
      ppr, pretty
@@ -50,7 +50,7 @@ import Language.Pads.Syntax
 import Language.Pads.BaseTypes
 import Language.Pads.Pretty
 import Language.Pads.PadsPrinter
-import Language.Pads.DataGen.Generation
+import Language.Pads.Generation
 
 import Data.Data
 import Data.List
